requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=6.0
scipy>=1.7.0
